Структуры UserResult и ProfileImageURL  вынес в отдельный файл ProfileImageModels
