+++ Image Feed/Controllers/SplashViewController.swift
@@ -106,7 +106,7 @@ final class SplashViewController: UIViewController, AuthViewControllerDelegate {
                 )
                 print("SplashViewController: Получен профиль для пользователя: \(profile.username)") // Лог получения профиля
                 
-                ProfileImageService.shared.fetchProfileImageURL(username: profileResult.username, in: self) { imageResult in
+                ProfileImageModels.shared.fetchProfileImageURL(username: profileResult.username, in: self) { imageResult in
                     switch imageResult {
                     case .success(let url):
                         print("SplashViewController: URL изображения профиля: \(url)") // Лог успешного получения URL
