+++ Image Feed/Profile/ProfileImageService.swift
@@ -24,13 +24,8 @@ final class ProfileImageService {
     // MARK: - Public Properties
     static let didChangeNotification = Notification.Name(rawValue: "ProfileImageService.didChangeNotification")
     
-    // MARK: - Models
-    struct UserResult: Codable {
-        let profileImage: [String: String]?
-    }
-    
-    // MARK: - Private Methods
-    private func makeAvatarRequest(username: String) -> URLRequest? {
+    // MARK: - Public Methods
+  func makeAvatarRequest(username: String) -> URLRequest? {
         guard let url = URL(string: "users/\(username)", relativeTo: Constants.defaultBaseURL)
         else {
             print ("ProfileService: неправильный URL")
@@ -48,39 +43,5 @@ final class ProfileImageService {
         print("ProfileImageService: Создан запрос с токеном: \(token)") // Лог ошибок
         return request
     }
-    
-    // MARK: - Public Methods
-    func fetchProfileImageURL(username: String, in viewController: UIViewController, _ completion: @escaping (Result<String, Error>) -> Void) {
-        
-        guard let request = makeAvatarRequest(username: username) else {
-            print("ProfileImageService: Невозможно создать запрос.")
-            completion(.failure(NetworkError.urlSessionError))
-            return
-        }
-        let task = urlSession.objectTask(for: request) { [weak self] (result: Result<UserResult, Error>) in
-            guard let self = self else { return }
-            
-            switch result {
-            case .success(let userResult):
-                if let profileImage = userResult.profileImage?["large"] {
-                    self.avatarURL = profileImage
-                    completion(.success(profileImage))
-                    print("ProfileImageService: Успешно получен URL для аватарки: \(profileImage)") // Лог ошибок
-                    NotificationCenter.default.post(
-                            name: ProfileImageService.didChangeNotification,
-                            object: self,
-                            userInfo: ["URL": profileImage])
-                } else {
-                    print("ProfileImageService: URL изображения профиля отсутствует.")
-                    completion(.failure(NetworkError.noData)) // Лог ошибок
-                }
-            case .failure(let error):
-                print("ProfileImageService: Ошибка при получении изображения профиля: \(error.localizedDescription)") // Лог ошибок
-                print("result: \(result)")
-                completion(.failure(error))
-            }
-        }
-        task.resume()
-    }
-    
+
 }
